Type Definition linux_raw_sys::ctypes::c_size_t

pub type c_size_t = usize;

🔬 This is a nightly-only experimental API. (c_size_t)

Equivalent to C’s size_t type, from stddef.h (or cstddef for C++).

This type is currently always usize, however in the future there may be platforms where this is not the case.