linux_raw_sys

Module elf

The ELF ABI. 🧝

This module is not as comprehensive as bindgened elf_uapi and provides only types for target pointer width: instead of elf32_phdr and elf64_phdr there’s only Elf_Phdr.

Structs

Elf_Dyn

Elf_Ehdr

Elf_Phdr

Elf_Rel

Elf_Rela

Elf_Sym

Elf_Verdaux

Elf_Verdef

Elf_auxv_t

Constants

DT_GNU_HASH

DT_HASH

DT_NULL

DT_REL

DT_RELA

DT_RELAENT

DT_RELASZ

DT_RELENT

DT_RELSZ

DT_STRTAB

DT_SYMENT

DT_SYMTAB

DT_VERDEF

DT_VERSYM

EI_ABIVERSION

EI_CLASS

EI_DATA

EI_NIDENT

EI_OSABI

EI_VERSION

ELFABIVERSION

ELFCLASS

ELFDATA

ELFMAG

ELFOSABI_LINUX

ELFOSABI_SYSV

EM_CURRENT

ET_DYN

EV_CURRENT

PF_R

PF_W

PF_X

PN_XNUM

PT_DYNAMIC

PT_GNU_RELRO

PT_GNU_STACK

PT_INTERP

PT_LOAD

PT_PHDR

PT_TLS

R_RELATIVE

SELFMAG

SHN_ABS

SHN_UNDEF

STB_GLOBAL

STB_WEAK

STN_UNDEF

STT_FUNC

STT_NOTYPE

STV_DEFAULT

VER_DEF_CURRENT

VER_FLG_BASE

Functions

ELF_ST_BIND

ELF_ST_TYPE

ELF_ST_VISIBILITY

Unions

Elf_Dyn_Union