linux_raw_sys::signal_macros

Function sig_ign

pub const fn sig_ign() -> __kernel_sighandler_t

Rust doesn’t currently permit us to use transmute to convert the SIG_IGN value into a function pointer in a const initializer, so we make it a function instead.